consolidate common code for change sources
