+++ packages/zero-cache/src/services/change-source/replica-schema.ts
@@ -1,0 +1,72 @@
+import type {LogContext} from '@rocicorp/logger';
+import type {Database} from '../../../../zqlite/src/db.ts';
+import {
+  runSchemaMigrations,
+  type IncrementalMigrationMap,
+  type Migration,
+} from '../../db/migration-lite.ts';
+import {AutoResetSignal} from '../change-streamer/schema/tables.ts';
+import {
+  CREATE_RUNTIME_EVENTS_TABLE,
+  recordEvent,
+} from '../replicator/schema/replication-state.ts';
+
+export async function initReplica(
+  log: LogContext,
+  debugName: string,
+  dbPath: string,
+  initialSync: (lc: LogContext, tx: Database) => Promise<void>,
+): Promise<void> {
+  const setupMigration: Migration = {
+    migrateSchema: (log, tx) => initialSync(log, tx),
+    minSafeVersion: 1,
+  };
+
+  await runSchemaMigrations(
+    log,
+    debugName,
+    dbPath,
+    setupMigration,
+    schemaVersionMigrationMap,
+  );
+}
+
+export async function upgradeReplica(
+  log: LogContext,
+  debugName: string,
+  dbPath: string,
+) {
+  await runSchemaMigrations(
+    log,
+    debugName,
+    dbPath,
+    // setupMigration should never be invoked
+    {
+      migrateSchema: () => {
+        throw new Error(
+          'This should only be called for already synced replicas',
+        );
+      },
+    },
+    schemaVersionMigrationMap,
+  );
+}
+
+export const schemaVersionMigrationMap: IncrementalMigrationMap = {
+  // There's no incremental migration from v1. Just reset the replica.
+  4: {
+    migrateSchema: () => {
+      throw new AutoResetSignal('upgrading replica to new schema');
+    },
+    minSafeVersion: 3,
+  },
+
+  5: {
+    migrateSchema: (_, db) => {
+      db.exec(CREATE_RUNTIME_EVENTS_TABLE);
+    },
+    migrateData: (_, db) => {
+      recordEvent(db, 'upgrade');
+    },
+  },
+};
